Move all logic in ts file for update-vacancies into new directive
